requirements.txt
alembic==1.8.1
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
decorator==4.4.2
Deprecated==1.2.13
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
idna==3.4
imageio==2.22.0
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
moviepy==1.0.3
mysqlclient==2.1.1
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
proglog==0.1.10
pyparsing==3.0.9
python-dotenv==0.21.0
pytube==12.1.0
redis==4.3.4
requests==2.28.1
rq==1.11.1
six==1.16.0
SQLAlchemy==1.4.41
tqdm==4.64.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1